movercomponent base class
